// src/app/Interfaces/assetHistory.ts
export class AssetHistory {
    AssetName: string;
    FacilityName: string;
    ZoneName: string;
    EntryTime: string;
    ExitTime: string;
    Status: string;

    Duration: number;

    Duration_years: number;

    Duration_months: number;

    Duration_days: number;

    Duration_hours: number;

    Duration_minutes: number;
    Duration_seconds: number;
}